Move methods back into mk_sse
